[refactor/#147] UseCase 파일 분리
