+++ app/src/main/java/com/poti/android/domain/usecase/auth/ObserveAuthStateUseCase.kt
@@ -1,0 +1,12 @@
+package com.poti.android.domain.usecase.auth
+
+import com.poti.android.domain.model.auth.AuthState
+import com.poti.android.domain.repository.AuthRepository
+import kotlinx.coroutines.flow.Flow
+import javax.inject.Inject
+
+class ObserveAuthStateUseCase @Inject constructor(
+    private val authRepository: AuthRepository,
+) {
+    operator fun invoke(): Flow<AuthState> = authRepository.observeAuthState()
+}
