+++ app/src/main/java/com/poti/android/domain/usecase/history/GetMyParticipationListUseCase.kt
@@ -1,0 +1,12 @@
+package com.poti.android.domain.usecase.history
+
+import com.poti.android.domain.model.history.MyPartyList
+import com.poti.android.domain.repository.ParticipationRepository
+import javax.inject.Inject
+
+class GetMyParticipationListUseCase @Inject constructor(
+    private val participationRepository: ParticipationRepository,
+) {
+    suspend operator fun invoke(status: String): Result<MyPartyList> =
+        participationRepository.getMyParticipationList(status)
+}
