+++ app/src/main/java/com/poti/android/domain/usecase/history/GetRecruitParticipantsUseCase.kt
@@ -1,0 +1,12 @@
+package com.poti.android.domain.usecase.history
+
+import com.poti.android.domain.model.history.ParticipantManageDetail
+import com.poti.android.domain.repository.PartyRepository
+import javax.inject.Inject
+
+class GetRecruitParticipantsUseCase @Inject constructor(
+    private val partyRepository: PartyRepository,
+) {
+    suspend operator fun invoke(postId: Long): Result<ParticipantManageDetail> =
+        partyRepository.getRecruitPostParticipant(postId)
+}
