+++ app/src/main/java/com/poti/android/domain/usecase/history/SubmitReviewUseCase.kt
@@ -1,0 +1,13 @@
+package com.poti.android.domain.usecase.history
+
+import com.poti.android.domain.repository.ReviewRepository
+import javax.inject.Inject
+
+class SubmitReviewUseCase @Inject constructor(
+    private val reviewRepository: ReviewRepository,
+) {
+    suspend operator fun invoke(
+        transactionId: Long,
+        star: Int,
+    ): Result<Long> = reviewRepository.postReview(transactionId, star)
+}
