--- app/src/main/java/com/poti/android/domain/usecase/party/PartyUseCases.kt
@@ -1,49 +1,0 @@
-package com.poti.android.domain.usecase.party
-
-import com.poti.android.domain.model.party.PartyDetail
-import com.poti.android.domain.model.party.PartyJoinInfo
-import com.poti.android.domain.model.party.PartyJoinOption
-import com.poti.android.domain.model.party.ProductPartyList
-import com.poti.android.domain.repository.PartyRepository
-import javax.inject.Inject
-
-class GetPartyDetailUseCase @Inject constructor(
-    private val partyRepository: PartyRepository,
-) {
-    suspend operator fun invoke(partyId: Long): Result<PartyDetail> =
-        partyRepository.getPartyDetail(partyId)
-}
-
-class GetPartyJoinOptionsUseCase @Inject constructor(
-    private val partyRepository: PartyRepository,
-) {
-    suspend operator fun invoke(partyId: Long): Result<PartyJoinOption> =
-        partyRepository.getPartyJoinOptions(partyId)
-}
-
-class JoinPartyUseCase @Inject constructor(
-    private val partyRepository: PartyRepository,
-) {
-    suspend operator fun invoke(joinInfo: PartyJoinInfo): Result<Long> =
-        partyRepository.postPartyJoin(joinInfo)
-}
-
-class GetProductPartyListUseCase @Inject constructor(
-    private val partyRepository: PartyRepository,
-) {
-    suspend operator fun invoke(
-        page: Int?,
-        size: Int?,
-        title: String,
-        artistId: Long,
-        sort: String,
-        memberIds: List<Long>?,
-    ): Result<ProductPartyList> = partyRepository.getProductPartyList(
-        page,
-        size,
-        title,
-        artistId,
-        sort,
-        memberIds,
-    )
-}
